backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
edge-tts
sqlalchemy
python-jose[cryptography]
bcrypt
python-multipart